# Conflicts:
#	src/main/java/com/example/JobApp/configuration/SecurityConfig.java

+++ src/main/java/com/example/JobApp/repository/UserRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface UserRepo extends JpaRepository<User,Integer> {
+public interface UserRepo extends JpaRepository<User,Long> {
     User findByUsername (String username);
 }
